+++ {{cookiecutter.project_slug}}/include/{{cookiecutter.project_slug}}/example.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
-#include <iostream>
+#include "{{cookiecutter.project_slug}}/exampleConfig.hpp"
 
-bool example() {
-  std::cout << "example thing!" << std::endl;
-  return true;
-}
+void SetupProjectConfig({{ cookiecutter.project_name.replace(' ', '') }}Config & config);
+void SetupProjectConfig({{ cookiecutter.project_name.replace(' ', '') }}Config & config, int argc, char* argv[]);
+bool example();
