+++ {{cookiecutter.project_slug}}/source/native.cpp
@@ -6,16 +6,20 @@
 
 #include "emp/base/vector.hpp"
 #include "emp/config/command_line.hpp"
+#include "emp/config/ArgManager.hpp"
 
 #include "{{cookiecutter.project_slug}}/example.hpp"
+#include "{{cookiecutter.project_slug}}/ProjectConfig.hpp"
+
+{{ cookiecutter.project_name.replace(' ', '') }}Config cfg;
 
 // This is the main function for the NATIVE version of {{cookiecutter.project_name}}.
 
 int main(int argc, char* argv[])
 {
-  emp::vector<std::string> args = emp::cl::args_to_strings(argc, argv);
-
   std::cout << "Hello, world!" << std::endl;
+  SetupProjectConfig(cfg, argc, argv);
+  cfg.Write(std::cout);
 
   return example();
 }
